django_app/requirements.txt
Django>=4.2.14, <4.3
djangorestframework>=3.15.2, <3.16
python-dotenv==1.0.1
django-unfold==0.34.0
gunicorn==23.0.0
psycopg2-binary==2.9.9